+++ test/argon2_test.exs
@@ -7,7 +7,7 @@ defmodule Argon2Test do
   test "only output encoded hash" do
     result = Base.hash_password("password", "somesalt")
     assert is_binary(result)
-    assert String.starts_with?(result, "$argon2")
+    assert String.starts_with?(result, "$argon2i$v=19$m=65536,t=6,p=1")
   end
 
   test "only output raw hash" do
@@ -21,7 +21,16 @@ defmodule Argon2Test do
     assert is_binary(raw)
     assert is_binary(encoded)
     refute String.starts_with?(raw, "$argon2")
-    assert String.starts_with?(encoded, "$argon2")
+    assert String.starts_with?(encoded, "$argon2i$v=19$m=65536,t=6,p=1")
+  end
+
+  test "customizing parameters with config" do
+    Application.put_env(:argon2_elixir, :t_cost, 3)
+    Application.put_env(:argon2_elixir, :m_cost, 12)
+    result = Base.hash_password("password", "somesalt")
+    assert String.starts_with?(result, "$argon2i$v=19$m=4096,t=3,p=1")
+    Application.delete_env(:argon2_elixir, :t_cost)
+    Application.delete_env(:argon2_elixir, :m_cost)
   end
 
   test "hashing and checking passwords" do